% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecomposeR.R
\name{DecomposeR}
\alias{DecomposeR}
\title{DecomposeR: Empirical Mode Decomposition for Cyclostratigraphy}
\description{
This package provides tools to apply Ensemble Empirical Mode
Decomposition (EEMD) for cyclostratigraphy purposes. It proposes a new
algorithm, that performs EEMD in seconds, a linear interpolation algorithm
using the greatest rational common divisor of depth or time, different
algorithms to compute instantaneous amplitude, frequency and ratios of
frequencies, and functions to verify and visualise the outputs.
}
\details{
Package: DecomposeR

Type: R package

Version: 1.0.7 (Summer 2025)

License: GPL-3
}
\note{
If you want to use this package for publication or research
purposes, please cite Wouters, S., Crucifix, M., Sinnesael, M., Da Silva,
A.C., Zeeden, C., Zivanovic, M., Boulvain, F., Devleeschouwer, X., 2022,
"A decomposition approach to cyclostratigraphic signal processing".
Earth-Science Reviews 225 (103894).
<doi:10.1016/j.earscirev.2021.103894>.
}
\author{
Sebastien Wouters

Maintainer: Sebastien Wouters \email{wouterseb@gmail.com}
}
