% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratios.R
\name{ratios}
\alias{ratios}
\title{Computes ratios of numerical values}
\usage{
ratios(x)
}
\arguments{
\item{x}{values to compute the ratio from}
}
\value{
a dataframe of $ratio, $x1 and $x2
}
\description{
Computes ratios of numerical values
}
\examples{
ratios(c(20,40,100,400))

}
